feat: Use proxy wiring for future event buffer (#12103)
