+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/event/FutureEventBuffer.java
@@ -16,21 +16,12 @@
 
 package com.swirlds.platform.event;
 
-import static com.swirlds.platform.consensus.ConsensusConstants.ROUND_FIRST;
-
-import com.swirlds.common.context.PlatformContext;
-import com.swirlds.common.metrics.FunctionGauge;
-import com.swirlds.common.sequence.map.SequenceMap;
-import com.swirlds.common.sequence.map.StandardSequenceMap;
+import com.swirlds.common.wiring.component.InputWireLabel;
 import com.swirlds.platform.consensus.NonAncientEventWindow;
-import com.swirlds.platform.eventhandling.EventConfig;
+import com.swirlds.platform.wiring.ClearTrigger;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import edu.umd.cs.findbugs.annotations.Nullable;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
-import java.util.concurrent.atomic.AtomicLong;
-import java.util.function.Function;
 
 /**
  * Buffers events from the future (i.e. events with a birth round that is greater than the round that consensus is
@@ -41,42 +32,7 @@ import java.util.function.Function;
  * Output from the future event buffer is guaranteed to preserve topological ordering, as long as the input to the
  * buffer is topologically ordered.
  */
-public class FutureEventBuffer {
-
-    /**
-     * A little lambda that builds a new array list. Cache this here so we don't have to create a new lambda each time
-     * we buffer a future event.
-     */
-    private static final Function<Long, List<GossipEvent>> BUILD_LIST = x -> new ArrayList<>();
-
-    private NonAncientEventWindow eventWindow;
-
-    private final SequenceMap<Long /* birth round */, List<GossipEvent>> futureEvents =
-            new StandardSequenceMap<>(ROUND_FIRST, 8, true, x -> x);
-
-    private final AtomicLong bufferedEventCount = new AtomicLong(0);
-
-    /**
-     * Constructor.
-     *
-     * @param platformContext the platform context
-     */
-    public FutureEventBuffer(@NonNull final PlatformContext platformContext) {
-        final AncientMode ancientMode = platformContext
-                .getConfiguration()
-                .getConfigData(EventConfig.class)
-                .getAncientMode();
-
-        eventWindow = NonAncientEventWindow.getGenesisNonAncientEventWindow(ancientMode);
-
-        platformContext
-                .getMetrics()
-                .getOrCreate(
-                        new FunctionGauge.Config<>("platform", "futureEventBuffer", Long.class, bufferedEventCount::get)
-                                .withDescription("the number of events sitting in the future event buffer")
-                                .withUnit("count"));
-    }
-
+public interface FutureEventBuffer {
     /**
      * Add an event to the future event buffer.
      *
@@ -84,23 +40,9 @@ public class FutureEventBuffer {
      * @return a list containing the event if it is not a time traveler, or null if the event is from the future and
      * needs to be buffered.
      */
+    @InputWireLabel("preconsensus event")
     @Nullable
-    public List<GossipEvent> addEvent(@NonNull final GossipEvent event) {
-        if (eventWindow.isAncient(event)) {
-            // we can safely ignore ancient events
-            return null;
-        } else if (event.getHashedData().getBirthRound() <= eventWindow.getPendingConsensusRound()) {
-            // this is not a future event, no need to buffer it
-            return List.of(event);
-        }
-
-        // this is a future event, buffer it
-        futureEvents
-                .computeIfAbsent(event.getHashedData().getBirthRound(), BUILD_LIST)
-                .add(event);
-        bufferedEventCount.incrementAndGet();
-        return null;
-    }
+    List<GossipEvent> addEvent(@NonNull GossipEvent event);
 
     /**
      * Update the current event window. As the event window advances, time catches up to time travelers, and events that
@@ -109,31 +51,15 @@ public class FutureEventBuffer {
      * @param eventWindow the new event window
      * @return a list of events that were previously from the future but are now from the present
      */
-    public List<GossipEvent> updateEventWindow(@NonNull final NonAncientEventWindow eventWindow) {
-        this.eventWindow = Objects.requireNonNull(eventWindow);
-
-        // We want to release all events with birth rounds less than or equal to the pending consensus round.
-        // In order to do that, we tell the sequence map to shift its window to the oldest round that we want
-        // to keep within the buffer.
-        final long oldestRoundToBuffer = eventWindow.getPendingConsensusRound() + 1;
-
-        final List<GossipEvent> events = new ArrayList<>();
-        futureEvents.shiftWindow(oldestRoundToBuffer, (round, roundEvents) -> {
-            for (final GossipEvent event : roundEvents) {
-                if (!eventWindow.isAncient(event)) {
-                    events.add(event);
-                }
-            }
-        });
-
-        bufferedEventCount.addAndGet(-events.size());
-        return events;
-    }
+    @InputWireLabel("non-ancient event window")
+    @Nullable
+    List<GossipEvent> updateEventWindow(@NonNull NonAncientEventWindow eventWindow);
 
     /**
      * Clear all data from the future event buffer.
+     *
+     * @param clearTrigger placeholder clearTrigger object
      */
-    public void clear() {
-        futureEvents.clear();
-    }
+    @InputWireLabel("clear")
+    void clear(@NonNull ClearTrigger clearTrigger);
 }
